webhooks/middleware/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
pillow==10.4.0
Flask==3.0.3
zeep==4.1.0
